Remove unused helper components and refactor imports for improved organization. Update Capabilities, FeaturedWork, Metrics, and Timeline sections to reference new paths for their respective containers. Clean up context files to align with the new structure.
