+++ src/components/capabilities/capabilitiesCard.tsx
