+++ src/components/capabilities/capabilitiesContainer.tsx
