+++ src/components/metrics/metricBar.tsx
